Integrate root.scala in Capability.scala
